Backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pymongo==4.10.1
minio==7.2.8
python-multipart==0.0.9
python-dotenv==1.1.1